refactor(notion): move canonical body fn into notion-effect-client (Option 2)
Relocate `canonicalizeBlockMarkdown` + the remark/unified/unist deps down from
`@overeng/notion-md` into `@overeng/notion-effect-client`, beside the renderer
(`treeToMarkdown`) and the media-URL canonicalizer it already calls. The move is
import-direction-legal (notion-effect-client does not depend on notion-md, no
cycle) and removes the dep duplication (remark leaves notion-md entirely).

`observeFromSnapshots` now canonicalizes the rendered body once, at the source,
before it flows into the inventory, the fidelity classifier, and the evidence
fingerprint — so the evidence, the classifier, pull, hash, and push all see the
same canonical bytes ("agree by construction", decision 0018). The pull choke
point (`remoteMarkdownFromBodyObservation`) therefore projects an already-
canonical body through verbatim instead of re-canonicalizing it.

`semanticEquivalent` stays in notion-md (it is sync policy, not the wire form)
and re-imports the canonical function. package.json regenerated via genie;
body-fidelity/body-observation re-baselined for the trailing-newline canonical
form (verdicts unchanged).

Refs the body/markdown pipeline consolidation (decision 0018).
