+++ packages/@overeng/notion-md/package.json
@@ -30,12 +30,7 @@
     "@overeng/notion-effect-client": "workspace:^",
     "@overeng/notion-effect-schema": "workspace:^",
     "@overeng/otel-contract": "workspace:^",
-    "@overeng/utils": "workspace:^",
-    "remark-gfm": "4.0.1",
-    "remark-parse": "11.0.0",
-    "remark-stringify": "11.0.0",
-    "unified": "11.0.5",
-    "unist-util-visit": "5.1.0"
+    "@overeng/utils": "workspace:^"
   },
   "devDependencies": {
     "@effect-atom/atom": "0.5.3",
